long int

+++ functions_nested_loops/102-fibonacci.c
@@ -8,13 +8,13 @@
 int main(void)
 {
 int i;
-int t1 = 1, t2 = 2;
-int next = t1 + t2;
+long t1 = 1, t2 = 2;
+long next = t1 + t2;
 
 printf("%d, %d, ", t1, t2);
 for(i = 3; 1 <= 50; ++i)
 {
-printf("%d, ", next);
+printf("%ld, ", next);
 t1 = t2;
 t2 = next;
 next = t1 + t2;
